tests: align aptos local and ci coverage
